% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_Ij.R
\name{E_Ij}
\alias{E_Ij}
\title{Expected item information}
\usage{
E_Ij(alpha_j, tau_j, mo_min, mo_max, distribution_table, ...)
}
\arguments{
\item{alpha_j}{- (numeric) discrimination parameter estimate}

\item{tau_j}{- (numeric) difficulty parameter estimate}

\item{mo_min}{- (integer) Lower bound of the age bracket (in months)}

\item{mo_max}{- (integer) Upperbound of the age bracket (in months)}

\item{distribution_table}{(data.frame) The tablue used to interpolate mean and SD estimates of abilities. The data frame requires the following named columns: (1) months, (2) mu, & (3) sigma.}

\item{...}{Additional arguments passed to simulate_abilities()}
}
\value{
A dataframe with an estimate of the expected information (est) along with the corresponding standard error (se).
}
\description{
This function calculates the expected information contributed by the the j-th item in a provided age interval. Note: Assumes logit link.
}
\examples{
E_Ij(alpha_j=1, tau_j=GCDG_est$tau[1], mo_min=0, mo_max=3, distribution_table=reference, N=5E4)
}
