% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_EI_table.R
\name{get_EI_table}
\alias{get_EI_table}
\title{Obtain expected item information table}
\usage{
get_EI_table(params_df, distribution_table, N = 1e+05, bin_width = 3,
  processors = NULL, seed = 4242)
}
\arguments{
\item{distribution_table}{(data.frame) The tablue used to interpolate mean and SD estimates of abilities. The data frame requires the following named columns: (1) months, (2) mu, & (3) sigma.}

\item{N}{- (integer) Number of draws for Monte Carlo integration.}

\item{bin_width}{- (integer) Width of bin in months. Must be a multiple of 36.}

\item{processors}{(integer) Number of processors for parallel computing. Defaults to total detected by parallel::detectCores().}

\item{seed}{(integer) Random number generating seed used in doRNG package during parallel processing.}
}
\value{
A data frame expected information statistic for each item (rows) and in each age bin (columns).
}
\description{
This function is a wrapper to obtain the expected item information statistics for each item and across age bins.
}
\examples{
get_EI_table(params_df = GCDG_est, distribution_table = reference, N = 1E6, bin_width=3, processors = 16, seed = 42)
}
