% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vartot_abilities.R
\name{get_vartot_abilities}
\alias{get_vartot_abilities}
\title{Total variability of abilities given an age interval}
\usage{
get_vartot_abilities(bin_width = 3, distribution_table, ...)
}
\arguments{
\item{bin_width}{- (integer) Width of bin in months. Must be a multiple of 36.}

\item{distribution_table}{(data.frame) The tablue used to interpolate mean and SD estimates of abilities. The data frame requires the following named columns: (1) months, (2) mu, & (3) sigma.}

\item{...}{Additional arguments passed to simulate_abilities()}
}
\value{
A dataframe with estimated values for the total variability of abilities in each age bin .
}
\description{
This function estimates the total variability of abilities across age bins of a specified interval width between 0-36 months.
}
\examples{
get_vartot_abilities(mo_min = 3, mo_max = 5.9999, distribution_table = reference)
}
