% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linterp_mu_sd.R
\name{linterp_mu_sd}
\alias{linterp_mu_sd}
\title{Interpolate means and SD of abilities}
\usage{
linterp_mu_sd(distribution_table, mo)
}
\arguments{
\item{distribution_table}{(data.frame) The tablue used to interpolate mean and SD estimates of abilities. The data frame requires the following named columns: (1) months, (2) mu, & (3) sigma.}

\item{mo}{(numeric vector) Months where interpolated mean and SD estimates will be interpolated. Must be between 0<=mo<=60.}
}
\value{
(data.frame) Interpolated means (mu) and standard deviations (sigma)
}
\description{
This function applies linear interpolation from a reference distribution table to obtain mean and standard devations of abilities at specified months of development.
}
\examples{
linterp_mu_sd(distribution_table = reference, mo = seq(0,36, len = 100))
}
