% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_abilities.R
\name{simulate_abilities}
\alias{simulate_abilities}
\title{Simulate abilities}
\usage{
simulate_abilities(mo_min, mo_max, distribution_table, N = 10000)
}
\arguments{
\item{mo_min}{- (integer) Lower bound of the age bracket (in months)}

\item{mo_max}{- (integer) Upper bound of the age bracket (in months)}

\item{distribution_table}{(data.frame) The tablue used to interpolate mean and SD estimates of abilities. The data frame requires the following named columns: (1) months, (2) mu, & (3) sigma.}

\item{N}{- (Integer) Number draws. Defaults to 10,000 draws.}
}
\value{
(numeric vector) Simulated ability values.
}
\description{
This function simulates abilities in a specified age interval assuming age is distributed uniformly.
}
\examples{
simulate_abilities(mo_min = 3, mo_max = 5.9999, distribution_table = reference)
}
